gpt-2-simple/requirements.txt
tensorflow-gpu
regex
requests
tqdm
numpy
toposort
pyyaml
googletrans